Merge branch 'main' of https://github.com/SCBI-ForestGEO/HarvardMortality

# Conflicts:
#	testthat/reports/daily_field_fix_error_counts.png
#	testthat/reports/map_of_error_and_warnings.png

+++ R_scripts/Generate_reports.R
@@ -208,7 +208,7 @@ idx_FAD <- !is.na(mort$FAD)
 idx_wound <- !is.na(mort$'Wounded main stem')
 idx_canker <- !is.na(mort$'Canker; swelling, deformity')
 idx_rot <- !is.na(mort$'Rotting trunk')
-idx_DWR <- !mort$'DWR' %in% "False"
+idx_DWR <- mort$'Dead with resprout' %in% "True"
 
 tag_stem_with_error <- paste(mort$Tag, mort$StemTag)[idx_trees & (idx_FAD | idx_wound | idx_wound | idx_canker | idx_rot | idx_DWR)]
 
@@ -226,7 +226,7 @@ idx_FAD <- !is.na(mort$FAD)
 idx_wound <- !is.na(mort$'Wounded main stem')
 idx_canker <- !is.na(mort$'Canker; swelling, deformity')
 idx_rot <- !is.na(mort$'Rotting trunk')
-idx_DWR <- !mort$'DWR' %in% "False"
+idx_DWR <- mort$'Dead with resprout' %in% "True"
 
 tag_stem_with_error <- paste(mort$Tag, mort$StemTag)[!idx_trees & (idx_FAD | idx_wound | idx_wound | idx_canker | idx_rot | idx_DWR)]
 
@@ -240,7 +240,7 @@ error_name <- "status_AU_but_DWR_selected"
 status_column <- rev(grep("Status", names(mort), value = T))[1]
 
 idx_trees <- mort[, status_column] %in% "AU"
-idx_DWR <- !mort$'DWR' %in% "False"
+idx_DWR <- mort$'Dead with resprout' %in% "True"
 
 tag_stem_with_error <- paste(mort$Tag, mort$StemTag)[idx_trees & idx_DWR]
 
@@ -442,7 +442,7 @@ if(!is.null(require_field_fix_error_file)) {
     file = file.path(here("testthat"), "reports/requires_field_fix/require_field_fix_error_file.csv"), 
     row.names = F
   )
-} else {
+} else if(file.exists(file.path(here("testthat"), "reports/requires_field_fix/require_field_fix_error_file.csv"))) {
   file.remove(file.path(here("testthat"), "reports/requires_field_fix/require_field_fix_error_file.csv"))
 }
 
@@ -452,7 +452,7 @@ if(!is.null(will_auto_fix_error_file)) {
     file = file.path(here("testthat"), "reports/will_auto_fix/will_auto_fix_error_file.csv"), 
     row.names = F
   )
-} else {
+} else if(file.exists(file.path(here("testthat"), "reports/will_auto_fix/will_auto_fix_error_file.csv"))) {
   file.remove(file.path(here("testthat"), "reports/will_auto_fix/will_auto_fix_error_file.csv"))
 }
 
@@ -462,7 +462,7 @@ if(!is.null(warning_file)) {
     file = file.path(here("testthat"), "reports/warnings/warnings_file.csv"), 
     row.names = F
   )
-} else {
+} else if(file.exists(file.path(here("testthat"), "reports/warnings/warnings_file.csv"))) {
   file.remove(file.path(here("testthat"), "reports/warnings/warnings_file.csv"))
 }
